packages/backend/services/ai-gateway/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic-settings==2.3.4
cryptography==42.0.8
